+++ frontend/ic_vetkeys/src/utils/utils.test.ts
@@ -10,6 +10,7 @@ import {
     augmentedHashToG1,
     deriveSymmetricKey,
     hashToScalar,
+    isValidTransportPublicKey,
     verifyBlsSignature,
 } from "./utils";
 import { expect, test } from "vitest";
@@ -38,6 +39,26 @@ test("creating random TransportSecretKey", () => {
 
     const pk = key.publicKeyBytes();
     assertEqual(pk.length, 48);
+});
+
+test("isValidTransportPublicKey", () => {
+    assertEqual(isValidTransportPublicKey(hexToBytes("50505050")), false);
+    assertEqual(
+        isValidTransportPublicKey(
+            hexToBytes(
+                "97f1d3a73197d7942695638c4fa9ac0fc3688c4f9774b905a14e3a3f171bac586c55e83ff97a1aeffb3af00adb22c6bb",
+            ),
+        ),
+        true,
+    );
+    assertEqual(
+        isValidTransportPublicKey(
+            hexToBytes(
+                "c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+            ),
+        ),
+        true,
+    );
 });
 
 test("parsing DerivedPublicKey", () => {